refactor: auto-split test_phase4_api.py
