+++ tests/api/phase4_api/test_a_i_p_handshake_response_contract.py
@@ -1,0 +1,97 @@
+"""Tests for Phase 4 API: Force overlays, actuator controls, model explorer, AIP.
+
+Validates Pydantic contract models and route logic for:
+- Force/torque vector overlays (#1199)
+- Per-actuator control sliders (#1198)
+- Model explorer & URDF editor (#1200)
+- AIP JSON-RPC server (#763)
+
+See issue #1199, #1198, #1200, #763
+"""
+
+from __future__ import annotations
+
+import pytest
+from pydantic import ValidationError
+from src.api.models.requests import (
+    ActuatorBatchCommandRequest,
+    ActuatorCommandRequest,
+    AIPJsonRpcRequest,
+    ForceOverlayRequest,
+    ModelCompareRequest,
+    ModelExplorerRequest,
+)
+from src.api.models.responses import (
+    ActuatorCommandResponse,
+    ActuatorInfo,
+    ActuatorPanelResponse,
+    AIPCapability,
+    AIPHandshakeResponse,
+    AIPJsonRpcResponse,
+    ForceOverlayResponse,
+    ForceVector3D,
+    ModelCompareResponse,
+    ModelExplorerResponse,
+    URDFTreeNode,
+)
+
+# ──────────────────────────────────────────────────────────────
+#  Contract Tests: Force Overlay (#1199)
+# ──────────────────────────────────────────────────────────────
+
+
+# ──────────────────────────────────────────────────────────────
+#  Contract Tests: Actuator Controls (#1198)
+# ──────────────────────────────────────────────────────────────
+
+
+# ──────────────────────────────────────────────────────────────
+#  Contract Tests: Model Explorer (#1200)
+# ──────────────────────────────────────────────────────────────
+
+
+# ──────────────────────────────────────────────────────────────
+#  Contract Tests: AIP JSON-RPC (#763)
+# ──────────────────────────────────────────────────────────────
+
+
+class TestAIPHandshakeResponseContract:
+    """Validate AIPHandshakeResponse model."""
+
+    def test_handshake(self) -> None:
+        """Full handshake response."""
+        resp = AIPHandshakeResponse(
+            server_name="UpstreamDrift AIP Server",
+            protocol_version="2.0",
+            capabilities=[
+                AIPCapability(
+                    name="simulation",
+                    version="1.0",
+                    methods=["simulation.start", "simulation.stop"],
+                ),
+                AIPCapability(
+                    name="model",
+                    version="1.0",
+                    methods=["model.load", "model.query"],
+                ),
+            ],
+            supported_methods=[
+                "simulation.start",
+                "simulation.stop",
+                "model.load",
+                "model.query",
+            ],
+        )
+        assert resp.server_name == "UpstreamDrift AIP Server"
+        assert len(resp.capabilities) == 2
+        assert len(resp.supported_methods) == 4
+
+
+# ──────────────────────────────────────────────────────────────
+#  Unit Tests: AIP Dispatcher (#763)
+# ──────────────────────────────────────────────────────────────
+
+
+# ──────────────────────────────────────────────────────────────
+#  Unit Tests: AIP Methods (#763)
+# ──────────────────────────────────────────────────────────────
