+++ tests/api/phase4_api/test_model_compare_request_contract.py
@@ -1,0 +1,78 @@
+"""Tests for Phase 4 API: Force overlays, actuator controls, model explorer, AIP.
+
+Validates Pydantic contract models and route logic for:
+- Force/torque vector overlays (#1199)
+- Per-actuator control sliders (#1198)
+- Model explorer & URDF editor (#1200)
+- AIP JSON-RPC server (#763)
+
+See issue #1199, #1198, #1200, #763
+"""
+
+from __future__ import annotations
+
+import pytest
+from pydantic import ValidationError
+from src.api.models.requests import (
+    ActuatorBatchCommandRequest,
+    ActuatorCommandRequest,
+    AIPJsonRpcRequest,
+    ForceOverlayRequest,
+    ModelCompareRequest,
+    ModelExplorerRequest,
+)
+from src.api.models.responses import (
+    ActuatorCommandResponse,
+    ActuatorInfo,
+    ActuatorPanelResponse,
+    AIPCapability,
+    AIPHandshakeResponse,
+    AIPJsonRpcResponse,
+    ForceOverlayResponse,
+    ForceVector3D,
+    ModelCompareResponse,
+    ModelExplorerResponse,
+    URDFTreeNode,
+)
+
+# ──────────────────────────────────────────────────────────────
+#  Contract Tests: Force Overlay (#1199)
+# ──────────────────────────────────────────────────────────────
+
+
+# ──────────────────────────────────────────────────────────────
+#  Contract Tests: Actuator Controls (#1198)
+# ──────────────────────────────────────────────────────────────
+
+
+# ──────────────────────────────────────────────────────────────
+#  Contract Tests: Model Explorer (#1200)
+# ──────────────────────────────────────────────────────────────
+
+
+class TestModelCompareRequestContract:
+    """Validate ModelCompareRequest model."""
+
+    def test_compare_request(self) -> None:
+        """Two-model comparison request."""
+        req = ModelCompareRequest(
+            model_a_path="model_a.urdf",
+            model_b_path="model_b.urdf",
+        )
+        assert req.model_a_path == "model_a.urdf"
+        assert req.model_b_path == "model_b.urdf"
+
+
+# ──────────────────────────────────────────────────────────────
+#  Contract Tests: AIP JSON-RPC (#763)
+# ──────────────────────────────────────────────────────────────
+
+
+# ──────────────────────────────────────────────────────────────
+#  Unit Tests: AIP Dispatcher (#763)
+# ──────────────────────────────────────────────────────────────
+
+
+# ──────────────────────────────────────────────────────────────
+#  Unit Tests: AIP Methods (#763)
+# ──────────────────────────────────────────────────────────────
